fix: order reply folder and styles
