+++ js/packages/components/inputs/chat/reply/interface.ts
@@ -1,3 +1,9 @@
+import { ReplyTargetInteraction } from '@berty/redux/reducers/chatInputs.reducer'
+
 export interface ReplyMessageProps {
 	convPK: string
+}
+
+export interface ActiveReplyInteractionProps {
+	activeReplyInteraction: ReplyTargetInteraction | null
 }
