+++ js/packages/contexts/styles/index.tsx
@@ -1,5 +1,5 @@
 export { useStyles, StyleProvider } from './styles.context'
 export { AppDimensionsProvider } from '../app-dimensions.context'
-export type { ColorsTypes, CustomStyleHook } from './types'
+export type { ColorsTypes } from './types'
 export { defaultStylesDeclaration } from './map-declaration'
 export * from './helpers'
